requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
numpy==1.26.1
pytz==2023.3.post1
shapely==2.0.2
sqlparse==0.4.4
